requirements.txt
python-dotenv
httpx[http2]
truststore
langchain-core
langchain-openai
langgraph
fastapi
uvicorn[standard]